Manual merge of old Tomcat8 plugin master branch into Tomcat plugin.
Previously there were two separate repos for Tomcat 7 (this repo) and Tomcat 8 (https://github.com/grails-plugins/grails-tomcat8-plugin). Going forward, the master branch in this repo represents Tomcat 8 and the 7.x branch is for Tomcat 7. This commit is a manual merge of the master branch from the old Tomcat8 repo into this repo.